# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "get": {
        "operationId": "CustomersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get all customers",
        "tags": [
          "customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create customer",
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get customer by ID",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Update customer",
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Delete customer",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/bookings/checkout": {
      "post": {
        "operationId": "BookingsController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Checkout cart and create booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/my": {
      "get": {
        "operationId": "BookingsController_findMyBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of customer bookings"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get current customer bookings",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/provider/{providerId}": {
      "get": {
        "operationId": "BookingsController_findByProvider",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get bookings assigned to a service provider",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings": {
      "get": {
        "operationId": "BookingsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All bookings returned"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get all bookings",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get booking details by ID",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}/cancel": {
      "patch": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Cancel booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/transactions/booking/{bookingId}": {
      "get": {
        "operationId": "TransactionsController_findByBooking",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction found"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction by booking ID — customer or super_user",
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/reviews/booking/{bookingId}": {
      "get": {
        "operationId": "ReviewsController_findByBooking",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List reviews for a booking",
        "tags": [
          "reviews"
        ]
      }
    },
...
    "/reviews/customer/{customerId}": {
      "get": {
        "operationId": "ReviewsController_findByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List reviews written by a customer",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "TatkuUnited API",
    "description": "Service platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste your JWT token (without the Bearer prefix)"
      },
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required by RolesGuard. Allowed values: super_user, region_manager, service_provider, customer"
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@mail.com"
          },
          "password": {
            "type": "string",
            "example": "Password@123"
          },
          "name": {
            "type": "string",
            "example": "Aditya Verma"
          },
          "phone": {
            "type": "string",
            "example": "9812345678"
          },
          "city": {
            "type": "string",
            "example": "Chennai"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "phone"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@mail.com"
          },
          "password": {
            "type": "string",
            "example": "Password@123"
          },
          "name": {
            "type": "string",
            "example": "Aditya Verma"
          },
          "phone": {
            "type": "string",
            "example": "9812345678"
          },
          "city": {
            "type": "string",
            "example": "Chennai"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "CheckoutBookingDto": {
        "type": "object",
        "properties": {
          "payment_method": {
            "type": "string",
            "description": "Payment method selected by the customer",
            "enum": [
              "UPI",
              "CARD",
              "NETBANK",
              "WALLET",
              "upi",
              "card",
              "netbank",
              "wallet"
            ],
            "example": "UPI"
          },
          "payment_gateway_ref": {
            "type": "string",
            "description": "Client or gateway payment reference, if already available",
            "example": "PGR20260503001"
          },
          "idempotency_key": {
            "type": "string",
            "description": "Client idempotency key, if already generated by the UI/payment layer",
            "example": "idem-checkout-20260503-001"
          }
        }
      },
...
    }
  }
}